Remove incorrect GiB multiplier
We make the config file *easy* by interpretting the input as GiB,
but Docker and the SolidFire cluster both work in a world of Bytes.

So what we have is a convoluted conversion process when we store
a default size in the Docker Driver and the Client structs.  We
were accidentally multiplying the settings twice, this change fixes
that.

In the future, we need to clean up the config file settings and var
names and make this much cleaner.