Implement usage of GetVolumeForAccount
Using just ListActiveVolumes is kinda nice and easy, BUT
it increases chances for overlapped Names and also results
in way more info in the log than we need.

Add GetVolumesForAccount and move things over to use that.

Also, be sure we check if the Volume already exists when
docker.daemon.driver.Create is called.