Fix client init for sfcli
sfcli creates an empy client var as part of the parsing
runbefore, then needs to actually set/replace with a real
valid client after reading the global flags.

We were using the form client := NewFromConfig, instead of just
setting it to the existing variable (client = NewFromConfig).

Fix that, so now the client can use the same config file as the
Docker Daemon