Fix list response in daemon driver
Calling `docker volume ls` wasn't showing any results for the
SolidFire driver.  Turns out the the problem was a simple issue
with the list filter using the wrong keyword.  This patch updates
daemon/driver.go to use the appropriate 'active' keyword so now
the response to the docker daemon actually includes volumes.

Addresses Github Issue: #1 and #2